refactor: login의 domain읆 auth에서 member로 변경
